+++ changelog.md
@@ -11,6 +11,7 @@
 - Added **Completed** and **Uncompleted** filters in the Custom Search Layer.
 - Added **Coin Rank** in the Difficulty Stats Popup.
 - Added **Custom Achievements**! Collect achievements by completing certain milestones and special interactions.
+- Added **Random Level** button in the Custom Search Layer to play a random rated layout level!
 - Fixed the Leaderboard and added refresh button to manually refresh the leaderboard data.
 - Added Indicators for Top 3 players in the Leaderboard Layer.
 - Change the sprites to be in a spritesheet.